requirements.txt
click==8.0.1
Flask==2.0.1
Flask-MySQLdb==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
Werkzeug==2.0.1
ConfigParser==5.3.0
Flask-Mobility==1.1.0
Flask-APScheduler==1.12.4
pythonping==1.1.4